Use Django's UserAdmin as base for custom UserAdmin.
